clue onvalue now in game status where its being used
