+++ src/components/RoomView/GameStatus.tsx
@@ -2,8 +2,10 @@ import React, { useEffect } from 'react';
 import { useSelector } from 'react-redux';
 import { useState } from 'react';
 import { RootState } from '../../store';
-import { child, update, ref } from 'firebase/database';
+import { child, update, ref, onValue } from 'firebase/database';
 import { database } from '../../utils/firebase';
+import { setCurrentClue } from '../../store/clueSlice';
+import { useDispatch } from 'react-redux';
 import './gameStatus.css';
 const GameStatus = () => {
   const { status, guessesRemaining, host } = useSelector((state: RootState) => state.game);
@@ -12,18 +14,20 @@ const GameStatus = () => {
   const { teamTwoOperatives, teamTwoSpymaster } = useSelector((state: RootState) => state.teamTwo);
   const { currentClue } = useSelector((state: RootState) => state.clues);
   const [playerNote, setPlayerNote] = useState<string>('');
+  const dispatch = useDispatch();
 
   // firebase references
   let playersInRoomRef = ref(database, `rooms/${roomId}/players/`);
   let hostRef = ref(database, `rooms/${roomId}/host`);
+  const currentClueRef = ref(database, `rooms/${roomId}/game/currentClue`);
 
   // possible game status strings
   const otherTeamSpymasterNote: string = 'The opponent spymaster is playing...wait for your turn';
   const sameTeamSpymasterNote: string = 'Wait for spymaster to give you a clue.';
   const spyGivingClueNote: string = 'Give your operatives a clue!';
-  const sameTeamOperativesNote: string = `THE CLUE IS ${currentClue.clueString}. Your operatives are guessing now`;
-  const otherTeamOperativesNote: string = `THE CLUE IS ${currentClue.clueString}. The opponent operative is playing...`;
-  const operativeGuessingNote: string = `THE CLUE IS ${currentClue.clueString}. Guess a word or click end turn`;
+  const sameTeamOperativesNote: string = `The clue is: ${currentClue.clueString}. Your operatives are guessing now`;
+  const otherTeamOperativesNote: string = `The clue is: ${currentClue.clueString}. The opponent operative is playing...`;
+  const operativeGuessingNote: string = `The clue is: ${currentClue.clueString}. Guess a word or click end turn`;
 
   const claimHost = () => {
     update(hostRef, { playerId, username });
@@ -134,6 +138,19 @@ const GameStatus = () => {
       }
     }
 
+  }, [status]);
+
+  useEffect(() => {
+    onValue(currentClueRef, (snapshot) => {
+      if (snapshot.exists()) {
+        const clue = snapshot.val();
+        dispatch(setCurrentClue(clue));
+      } else {
+        console.log('no clue exists')
+      }
+    });
+    /*game status needs to be watched because rendering the current clue 
+    depends on whose turn it is / what will show in gamestatus bar */
   }, [status]);
 
   const showGuessesRemaining = !(guessesRemaining === 0 || guessesRemaining === undefined);
