+++ src/components/RoomView/clue/Clue.tsx
@@ -3,17 +3,15 @@ import pluralize from 'pluralize';
 import { ClueType } from '../../../utils/interfaces'; // for Typescript interface
 // firebase:
 import { database } from '../../../utils/firebase';
-import { ref, child, push, update, off, set, onValue } from 'firebase/database';
+import { ref, child, push, update, off, set } from 'firebase/database';
 // redux:
 import { useSelector, useDispatch } from 'react-redux';
 import { setCurrentClue } from '../../../store/clueSlice';
 import { RootState } from '../../../store';
 // CSS:
 import './clue.css';
-import Stack from '@mui/material/Stack';
 import Snackbar from '@mui/material/Snackbar';
 import MuiAlert, { AlertProps } from '@mui/material/Alert';
-import { useEffect } from 'react';
 
 const Alert = React.forwardRef<HTMLDivElement, AlertProps>(function Alert(props, ref) {
   return <MuiAlert elevation={6} ref={ref} variant="filled" {...props} />;
@@ -50,7 +48,6 @@ const Clue = () => {
     if (reason === 'clickaway') {
       return;
     }
-
     setOpen(false);
   };
   //
@@ -138,19 +135,6 @@ const Clue = () => {
   const iAmSpy2AndItsMyTurn = gameStatus === 'team2SpyTurn' && teamTwoSpymaster?.playerId === playerId;
   // If either of these above are true, show the Clue component!
   if (iAmSpy1AndItsMyTurn || iAmSpy2AndItsMyTurn) showClue = true;
-
-  useEffect(() => {
-    onValue(currentClueRef, (snapshot) => {
-      if (snapshot.exists()) {
-        const clue = snapshot.val();
-        dispatch(setCurrentClue(clue));
-      } else {
-        console.log('no clue exists')
-      }
-    });
-    /*game status needs to be watched because rendering the current clue 
-    depends on whose turn it is / what will show in gamestatus bar */
-  }, [gameStatus]);
 
 
   if (!showClue) return <></>;
